Bug amb games solved

+++ Hidato/src/CapaPersistencia/GameDBController.java
@@ -80,12 +80,14 @@ public class GameDBController {
     
     public void deleteAllGames(String username) {
         File f = new File(directory + username);
-        ArrayList<String> gameNames = new ArrayList(Arrays.asList(f.list()));
-        for (int i = 0; i < gameNames.size(); ++i) {
-            File aux = new File(directory + username + "/" + gameNames.get(i));
-            aux.delete();
+        if (Arrays.asList(f.list()) != null) {
+            ArrayList<String> gameNames = new ArrayList(Arrays.asList(f.list()));
+            for (int i = 0; i < gameNames.size(); ++i) {
+                File aux = new File(directory + username + "/" + gameNames.get(i));
+                aux.delete();
+            }
+            f.delete();
         }
-        f.delete();
     }
